Remove File::open_path interface method in favor of File::open_stream
It was only a helper to remove the call to std::fs::File::open, so it didn't
really hide any meaningful complexity from the user. In doing so, though, it
forced the public interface to depend on std::fs and std::path, which gets in
the way of offering a no_std option (#19).