Rework ParseError to be more useful
This changes ParseError into a variated enum which allows users (and tests) to
easily match on specific error conditions.

It also encodes helpful associated data into the various enum variants in order
to provide more helpful error messages for the various error conditions.

This pattern of enum error construction also defers string formatting to the
user rather than forcing an expensive format! when the error is generated.

This also properly implements std::error::Error's .source() as requested in #15

This also exposes ParseError in the public interface as requested in #20