Get rid of unnecessary EndianParseExt trait
There was only one implementer of this trait, on &[u8], and the indirection
didn't serve any real purpose. The indirection via the double &&[u8] impl was
more expensive than necessary, and the indirection also added a layer of unnecessary
cognitive overhead to reading the code.

This goes back to having simple parse methods for the endian-aware integer parsing
which operate on byte slices.