Add a generic ParsingTable impl to go alongside the ParsingIterator
Also, use this ParsingTable to replace the hand-rolled SymbolTable.

This implements a lazily-parsed table/array of structs that can be parsed
with the ParseAt trait that drives the parsing of our ELF structs.

Elements in the table are lazily parsed on-demand via the .get(index)
method. The table can also optionally be converted into a ParsingIterator
with either iter() or into_iter(), meaning the ParsingTable can be used
like so:

```
let table = ParsingTable<Foo>::new(...)
for foo in Table {
    ... do things with parsed Foo
}
```

or
```
let table = ParsingTable<Foo>::new(...)
for foo in Table.iter() {
    ... do things with parsed Foo
}
for foo in Table.iter() {
    ... do another pass of things with parsed Foo
}
```

Note that the second pattern will re-parse the second time. Useful
if you're wanting to avoid allocations, but otherwise it's probably
more useful to .collect() the iterator into a Vec<Foo>.