Make ElfStream parse out the SectionHeaders into an allocated Vec as part of open_stream()
The shdrs are commonly used for nearly every other interface method, so there's no need to parse them
out lazily all the time. This interface can (and does) allocate, so let's just allocate and parse
them up front to get it out of the way for the other interface methods.