Merge the CHANGELOG + crate version bump for v0.7.3 onto main.
The v0.7.3 release was branched out in order to not include the changes on main to
parsed note name representation (&str -> &[u8]) found in d29e9964dc617 and
a853cb24bd6d.

That was done in order to not ship breaking API changes in a 0.7.X release, as
I want to bundle those into their own 0.8.X once I have a few more changes to
warrant the larger bump. And as always, if a friendly human is finding that
they want a published crate version with those changes, feel free to reach out
with an issue and I'll happily work with you to bundle those changes into a
published release.