Make to_str no_std compatible
This patch allows the use of the (applicable subset of the) to_str
module possible under no_std with or without the alloc crate.

To this end, a new feature flag for alloc is introduced (implied by
std). All use of String in to_str is gated behind the alloc feature.