Drop "nightly" feature in favor of stabilized core::error::Error for ParseError when building for no_std
The nightly error_in_core feature was stabilized in rust version 1.81.0, so we
can now remove the use of our custom "nightly" feature as a crutch to signal
its availability. We instead use it unconditionally when std feature is unavailable
and bump the crate's MSRV to 1.81.0.

Note that this could be an undesireable MSRV bump for folks consuming this crate,
particularly for those who aren't building without std. As such, I'm keeping
this on the 0.8.0-dev branch only and not adding this to the 0.7 branch.